Handle Hex String in Utils.ToDecimal
