+++ .changeset/lemon-ads-fix.md
@@ -2,4 +2,4 @@
 "chainlink": patch
 ---
 
-#added Handle Hex String in EA Telemetry
+#added Handle Hex String in Utils.ToDecimal
