+++ core/services/ocrcommon/telemetry.go
@@ -5,10 +5,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"math/big"
-	"strings"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/shopspring/decimal"
 	"github.com/smartcontractkit/libocr/commontypes"
 	ocrtypes "github.com/smartcontractkit/libocr/offchainreporting2plus/types"
 	"google.golang.org/protobuf/proto"
@@ -228,32 +226,12 @@ func getJsonParsedValue(trr pipeline.TaskRunResult, trrs *pipeline.TaskRunResult
 	if nextTask != nil && nextTask.Task.Type() == pipeline.TaskTypeJSONParse {
 		asDecimal, err := utils.ToDecimal(nextTask.Result.Value)
 		if err != nil {
-			v, ok := nextTask.Result.Value.(string)
-			if !ok {
-				return nil
-
-			}
-			hexAnswer, success := hexStringToDecimal(v)
-			if !success {
-				return nil
-			}
-			asDecimal = hexAnswer
+			return nil
 		}
 		toFloat, _ := asDecimal.Float64()
 		return &toFloat
 	}
 	return nil
-}
-
-// hexStringToDecimal takes in a hex string, and returns (decimal.Decimal, bool), bool indicates success status
-func hexStringToDecimal(hexString string) (decimal.Decimal, bool) {
-	hexString = strings.TrimPrefix(hexString, "0x")
-	n := new(big.Int)
-	_, success := n.SetString(hexString, 16)
-	if !success {
-		return decimal.Zero, false
-	}
-	return decimal.NewFromBigInt(n, 0), true
 }
 
 // getObservation checks pipeline.FinalResult and extracts the observation
